+++ src/client/java/com/carlschierig/rudimentary/RudimentaryClient.java
@@ -1,9 +1,11 @@
 package com.carlschierig.rudimentary;
 
+import com.carlschierig.rudimentary.ui.screen.RudimentaryScreens;
 import net.fabricmc.api.ClientModInitializer;
 
 public class RudimentaryClient implements ClientModInitializer {
-	@Override
-	public void onInitializeClient() {
-	}
+    @Override
+    public void onInitializeClient() {
+        RudimentaryScreens.init();
+    }
 }