+++ src/client/java/com/carlschierig/rudimentary/ui/screen/BrickFurnaceScreen.java
@@ -1,0 +1,18 @@
+package com.carlschierig.rudimentary.ui.screen;
+
+import com.carlschierig.rudimentary.menu.BrickFurnaceMenu;
+import net.minecraft.client.gui.screens.inventory.AbstractFurnaceScreen;
+import net.minecraft.client.gui.screens.recipebook.SmeltingRecipeBookComponent;
+import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.entity.player.Inventory;
+
+public class BrickFurnaceScreen extends AbstractFurnaceScreen<BrickFurnaceMenu> {
+    private static final ResourceLocation LIT_PROGRESS_SPRITE = ResourceLocation.withDefaultNamespace("container/furnace/lit_progress");
+    private static final ResourceLocation BURN_PROGRESS_SPRITE = ResourceLocation.withDefaultNamespace("container/furnace/burn_progress");
+    private static final ResourceLocation TEXTURE = ResourceLocation.withDefaultNamespace("textures/gui/container/furnace.png");
+
+    public BrickFurnaceScreen(BrickFurnaceMenu menu, Inventory inventory, Component component) {
+        super(menu, new SmeltingRecipeBookComponent(), inventory, component, TEXTURE, LIT_PROGRESS_SPRITE, BURN_PROGRESS_SPRITE);
+    }
+}