+++ src/main/java/com/carlschierig/rudimentary/util/TranslationUtil.java
@@ -1,0 +1,8 @@
+package com.carlschierig.rudimentary.util;
+
+public class TranslationUtil {
+
+    public static String translateContainer(String id) {
+        return "container.privileged." + id;
+    }
+}